+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Removed
 * LLVM backend because it was not used much and not good integrated in pystencils.
+* OpenCL backend because it was not used much and not good integrated in pystencils.